# openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "API-Projekt",
    "description": "Schulprojekt der IFA01, LF9 (Herr Wichmann) zur Erstellung einer Rest-API",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://127.0.0.1/"
    }
  ],
  "paths": {
    "/list/{list_id}": {
      "get": {
        "summary": "Gets a list",
        "description": "Returns all entries of a to-do-list",
        "parameters": [
          {
            "name": "list_id",
            "in": "path",
            "description": "ID of list",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoList"
                }
              }
            }
          },
          "404": {
            "description": "invalid list id",
            "content": {}
          }
        }
      }
    },
    "/list/": {
        "post" : {
          "summary": "Add a new list",
          "description": "Creates a new list to be filled with entries",
          "parameters": [
            {
              "name": "NewList",
              "in": "header",
              "description": "New List Object",
              "required": true,
              "schema": {
                "$ref":"#/components/schemas/TodoList"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TodoList"
                  }
                }
              }
            },
            "500": {
              "description": "internal server error",
              "content": {}
            }
          }
        }
      },
      "/list/{list_id}/entry":{
        "post":{
          "summary": "Adds a new list entry",
          "description": "Creates a new entry in a given list",
          "parameters": [
            {
              "name": "list_id",
              "in": "path",
              "description": "ID of list",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/TodoList"
                  }
                }
              }
            },
            "500": {
              "description": "internal server error",
              "content": {}
            }
          }
        }
      },
      "/list/{list_id}/entry/{entry_id}":{
        "delete" : {
          "summary": "Deletes an entry",
          "description": "Deletes an entry of a to-do-list",
          "parameters": [
            {
              "name": "list_id",
              "in": "path",
              "description": "ID of list",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "entry_id",
              "in": "path",
              "description": "ID of list",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema" : {
                     "type": "integer",
                     "format": "int32"
                  }
                }
              }
            },
            "404": {
              "description": "invalid list id",
              "content": {}
            }
          }
        },
        "post" : {
          "summary": "Updates an entry",
          "description": "Updates an existing entry",
          "parameters": [
            {
              "name": "list_id",
              "in": "path",
              "description": "ID of list",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            },
            {
              "name": "entry_id",
              "in": "path",
              "description": "ID of list",
              "required": true,
              "schema": {
                "type": "integer",
                "format": "int32"
              }
            }
          ],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema" : {
                     "type": "integer",
                     "format": "int32"
                  }
                }
              }
            },
            "404": {
              "description": "invalid list id",
              "content": {}
            }
          }  
        }
      },
      "/users/": {
        "get" : {
          "summary": "Shows all users",
          "description": "Returns a list of all users",
          "parameters": [],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/users"
                  }
                }
              }
            },
            "500": {
              "description": "internal server error",
              "content": {}
            }
          }
        }
      },
      "/user/": {
        "post" : {
          "summary": "Adds User",
          "description": "Adds a new User",
          "parameters": [{
            "name": "NewUser",
            "in": "header",
            "description": "Creates a new user object",
            "required": true,
            "schema": {
              "$ref":"#/components/schemas/user"
            }
          }],
          "responses": {
            "200": {
              "description": "successful operation",
              "content": {
                "application/json": {
                  "schema": {
                    "$ref": "#/components/schemas/users"
                  }
                }
              }
            },
            "500": {
              "description": "internal server error",
              "content": {}
            }
          }
        }
      },
      "/user/{user_id}":{
        "delete" : {
        "summary": "Deletes a user",
        "description": "Deletes a user via user ID",
        "parameters": [{
          "name": "user_id",
          "in": "path",
          "description": "ID of user",
          "required": true,
          "schema": {
            "type": "integer",
            "format": "int32"
          }    
        }],
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/users"
                }
              }
            }
          },
          "500": {
            "description": "internal server error",
            "content": {}
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "TodoList": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "entries": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TodoEntry"
            }
          }
        }
      },
      "TodoEntry": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "list_id": {
            "type": "string"
          },
          "owner_id": {
            "type": "string"
          }
        }
      },
      "users" : {
        "type" : "object",
        "properties" : {
          "user_list" :{
            "type": "array",
            "items": {
              "$ref":"#/components/schemas/user"
            }
          }
        }
      },
      "user" : {
         "type": "object",
         "properties" :{
            "id" : {
              "type":"string"
            },
            "name": {
              "type": "string"
            }
         }
      }
    }
  }
}
  